requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==23.1.0
click==8.1.3
Django==3.2
djangorestframework==3.12.4
flake8==6.0.0
flake8-commas==2.1.0
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.3.0
platformdirs==3.1.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
